Перенёс разбивку по дням в метод презентера
