+++ src/utils/common.js
@@ -24,32 +24,3 @@ export const createDateInAttributeFormat = (date) => {
   const day = date.getDate();
   return `${year}-${month}-${day}`;
 };
-
-export const generateDays = (events) => {
-  const eventsDates = events.map((event) => {
-    return event.dateStart;
-  });
-
-  eventsDates.sort((a, b) => a - b);
-
-  let days = new Set();
-
-  eventsDates.forEach((date) => {
-    days.add(date);
-  });
-
-  days = Array.from(days).map((day) => {
-    const eventsInDay = events.filter((event) => {
-      return event.dateStart === day;
-    });
-    eventsInDay.sort((a, b) => {
-      return Date.parse(a.dateTimeStart) - Date.parse(b.dateTimeStart);
-    });
-    return {
-      day,
-      events: eventsInDay,
-    };
-  });
-
-  return days;
-};
